% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_deleteNA_df.R
\name{aux_deleteNA_df}
\alias{aux_deleteNA_df}
\title{Removing NAs observations from a data frame}
\usage{
aux_deleteNA_df(data)
}
\arguments{
\item{data}{is a data.frame object}
}
\value{
a data.frame object with the same structure as data and no observation with missing (NA) values
}
\description{
Internal, auxillary functions
}
\examples{
df=data.frame(x=c(rnorm(10),NA,NA),y=c(NA,NA,rnorm(10)))
SLEMI:::aux_deleteNA_df(df)
}
\keyword{internal}
