% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_input_checks.R
\name{aux_input_checks}
\alias{aux_input_checks}
\title{Auxiliary function. Initial verification of input}
\usage{
aux_input_checks(data, signal, response, side_variables)
}
\arguments{
\item{data}{is an input object that should be a data.frame}

\item{signal}{is a character object that indicates input columns of data}

\item{response}{is a character vector that indicates output's columns of data}

\item{side_variables}{is a character vector that indicates side variables' columns of data}
}
\value{
If all initial data is valid, string "ok" is retured. Otherwise, error is given.
}
\description{
Auxiliary function. Initial verification of input
}
\examples{
data=data_example1
aux_input_checks(data=data,signal="signal",response="response",side_variables="sideVar")

data=as.matrix(data_example1)
aux_input_checks(data=data,signal="signal",response="response",side_variables="sideVar")

data=data_example1
aux_input_checks(data=data,signal="input",response="response",side_variables="sideVar")
}
\keyword{internal}
