% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_signal_transform.R
\name{aux_signal_transform}
\alias{aux_signal_transform}
\title{Auxiliary function. Initial verification of signal provided and its transformation into a factor type}
\usage{
aux_signal_transform(data, signal)
}
\arguments{
\item{data}{is a data.frame}

\item{signal}{is a character that indicates columns of data that include the labels of input}
}
\value{
A data.frame that is a copy of data provided with signal column transformed to factor class. 
If signal has been numeric initially, additional column is created "signal_RAW" that is an exact
copy of original column
}
\description{
Auxiliary function. Initial verification of signal provided and its transformation into a factor type
}
\examples{
data=data_example1
data1=aux_signal_transform(data,"signal")
data$signal=as.character(data$signal)
data2=aux_signal_transform(data,"signal")
data$signal=as.numeric(data$signal)
data3=aux_signal_transform(data,"signal")
}
\keyword{internal}
