% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_theme_publ.R
\name{aux_theme_publ}
\alias{aux_theme_publ}
\title{Tuning GGplot Themes}
\usage{
aux_theme_publ(base_size = 12, base_family = "sans", version = 1)
}
\arguments{
\item{base_size}{integer that sets the default size of font used in the plot}

\item{base_family}{character that indicates what type of font should be used}

\item{version}{integer that changes the characteristic of the plot, values 1,2 and 3 accepted.}
}
\description{
Internal, auxillary functions
}
\details{
This function changes the theme of plots created with the use of ggplot package
}
\examples{
library(ggplot2)
ggplot(data=data.frame(x=1:10,y=rnorm(10)),aes(x=x,y=y))+
geom_point()+SLEMI:::aux_theme_publ(version=2)
}
\keyword{internal}
