% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_x_log_y.R
\name{aux_x_log_y}
\alias{aux_x_log_y}
\title{Calculation of expression \eqn{x\cdot \log y}}
\usage{
aux_x_log_y(x, y)
}
\arguments{
\item{x}{is a numeric vector}

\item{y}{is a numeric vector (the same length of x)}
}
\value{
Function calculates the value of expression \eqn{x\cdot \log y} element-wise in a numerically stable way. 
The result is a numeric vector of the same length as x. It is assumed that \eqn{0\cdot \log 0 = 0}.
}
\description{
Internal, auxillary functions
}
\examples{
SLEMI:::aux_x_log_y(1,2)
SLEMI:::aux_x_log_y(0,0)
SLEMI:::aux_x_log_y(1000,100)
}
\keyword{internal}
