% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_klogreg_algorithm.R
\name{capacity_klogreg_algorithm}
\alias{capacity_klogreg_algorithm}
\title{Algorithm for channel capacity calculation using glmnet algorithm}
\usage{
capacity_klogreg_algorithm(dataMatrix, dataSignal, model_out = TRUE,
  cc_maxit = 100, lambda_num = 10, cv_core_num = 1)
}
\arguments{
\item{dataMatrix}{is a numeric matrix with columns treated as explanatory variables in 
logistic regression algorithm (only if glmnet_algorithm=TRUE)}

\item{dataSignal}{is a vector of factors indicationg channel's input}

\item{model_out}{is the logical indicating if the calculated logisitc regression model should be included in output list}

\item{cc_maxit}{is the number of iteration of iterative algorithm of finding maximum CC}

\item{lambda_num}{is the lambda parameter of glmnet package}

\item{cv_core_num}{is the number of cores to use in parallel computing of glmnet package}
}
\value{
a list with three elements:
\itemize{
\item output$regression - confusion matrix of logistic regression predictions
\item output$cc         - channel capacity in bits
\item output$p_opt      - optimal probability distribution
\item output$model_out      - glmnet object describing logistic regression model (if model_out=TRUE)
}
}
\description{
Algorithm for channel capacity calculation using glmnet algorithm
}
