% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_logreg_algorithm.R
\name{capacity_logreg_algorithm}
\alias{capacity_logreg_algorithm}
\title{Algorithm for channel capacity calculation}
\usage{
capacity_logreg_algorithm(data, signal = "signal", response = "response",
  side_variables = NULL, formula_string = NULL, model_out = TRUE,
  cc_maxit = 100, lr_maxit = 1000, MaxNWts = 5000)
}
\arguments{
\item{data}{must be a data.frame object}

\item{signal}{is a character object that indicates columns of data to be treated as channel's input}

\item{response}{is a character vector that indicates columns of data to be treated as channel's output}

\item{side_variables}{is an optional character vector that indicates side variables' columns of data, if NULL no side variables are included}

\item{formula_string}{is a character object that includes a formula syntax to use in algorithm for capacity calculation}

\item{model_out}{is the logical indicating if the calculated logisitc regression model should be included in output list}

\item{cc_maxit}{is the number of iteration of iterative algorithm of finding maximum CC}

\item{lr_maxit}{is the number of iteration of iterative algorithm of logistic regression}

\item{maxNWts}{is the maximum acceptable number of weights in logistic regression algorithm - the higher, the slower the computations}
}
\value{
a list with three elements:
\itemize{
\item output$regression - confusion matrix of logistic regression predictions
\item output$cc         - channel capacity in bits
\item output$p_opt      - optimal probability distribution
\item output$model      - nnet object describing logistic regression model (if model_out=TRUE)
}
}
\description{
INPUT:
}
