% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_logreg_testing.R
\name{capacity_logreg_testing}
\alias{capacity_logreg_testing}
\title{Testing procedures for calculating channel capacity}
\usage{
capacity_logreg_testing(data, signal = "signal", response = "response",
  side_variables = NULL, cc_maxit = 100, lr_maxit = 1000,
  MaxNWts = 5000, formula_string = NULL, glmnet_algorithm = FALSE,
  dataMatrix = NULL, glmnet_lambdanum = 10, model_out = FALSE,
  TestingSeed = 1234, testing_cores = 4, boot_num = 10, boot_prob = 0.8,
  sidevar_num = 10, traintest_num = 10, partition_trainfrac = 0.6)
}
\arguments{
\item{data}{must be a data.frame object}

\item{signal}{is a character object that indicates columns of data to be treated as channel's input}

\item{response}{is a character vector that indicates columns of data to be treated as channel's output}

\item{side_variables}{is an optional character vector that indicates side variables' columns of data, if NULL no side variables are included}

\item{cc_maxit}{is the number of iteration of iterative algorithm of finding maximum CC}

\item{lr_maxit}{is the number of iteration of iterative algorithm of logistic regression}

\item{formula_string}{is a character object that includes a formula syntax to use in algorithm for capacity calculation}

\item{glmnet_algorithm}{is the logical indicating if the glmnet package should be used}

\item{dataMatrix}{is a numeric matrix with columns treated as explanatory variables in 
logistic regression algorithm (only if glmnet_algorithm=TRUE)}

\item{glmnet_lambdanum}{is the lambda parameter of glmnet package}

\item{TestingSeed}{is the seed used in bootstrap procedures (only if testing=TRUE)}

\item{testing_cores}{- number of cores to use in parallel computing if implemented (only if testing=TRUE)}

\item{boot_num}{is the number of bootstrap tests to be performed (only if testing=TRUE)}

\item{boot_prob}{is the proportion of initial size of data to be used in bootstrap (only if testing=TRUE)}

\item{traintest_num}{is the number of traintest tests to be performed (only if testing=TRUE)}

\item{partition_trainfrac}{is the fraction of data to be used as a training dataset (only if testing=TRUE)}

\item{maxNWts}{is the maximum acceptable number of weights in logistic regression algorithm - the higher, the slower the computations}

\item{resamp_num}{is the number of resmapling tests to be performed (only if testing=TRUE)}

\item{glmnet_cores}{is the number of cores to use in parallel computing of glmnet package}
}
\value{
a list with four elements:
\itemize{
\item output$bootstrap - confusion matrix of logistic regression predictions
\item output$resamplingMorph          - channel capacity in bits
\item output$traintest      - optimal probability distribution
\item output$bootResampMorph      - nnet object describing logistic regression model (if model_out=TRUE)
}
Each of above is a list, where an element is an output of a single repetition of the channel capacity algorithm
}
\description{
INPUT:
}
\examples{

}
