% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapacityLogReg.R
\docType{package}
\name{CapacityLogReg}
\alias{CapacityLogReg}
\alias{CapacityLogReg-package}
\title{CapacityLogReg: A package for estimating channel capacity from experimental data}
\description{
The package CapacityLogReg is designed to estimate channel capacity between finite 
state input and multidimensional output from experimental data. For efficient computations, 
it uses iterative algorithm based on logistic regression.
}
