% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example2}
\alias{data_example2}
\title{Examplary data set II}
\format{A data frame with 1500 rows and 4 variables:
\describe{
\item{signal}{Label of input}
\item{X1}{The value of first dimension of output}
\item{X2}{The value of second dimension of output}
\item{X3}{The value of third dimension of output}
}}
\source{
synthetic
}
\usage{
data_example2
}
\description{
A dataset describing a channel with 3 possible inputs and 3-dimensional output
with 500 observations per input.
Conditional output distributions are multivariate Gaussians.
}
\keyword{datasets}
