% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteNA_df.R
\name{deleteNA_df}
\alias{deleteNA_df}
\title{Removing NAs observations from data frame}
\usage{
deleteNA_df(data)
}
\arguments{
\item{data}{is a data.frame object}
}
\value{
a data.frame object with the same structure as data and no observation with missing (NA) values
}
\description{
Removing NAs observations from data frame
}
\examples{

}
\keyword{internal}
