% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_generator.R
\name{formula_generator}
\alias{formula_generator}
\title{Formula for logistic regression algorithm}
\usage{
formula_generator(signal = "signal", response = "response",
  side_variables = NULL)
}
\arguments{
\item{signal}{is a character object that indicates columns of data to be treated as channel's input}

\item{response}{is a character vector that indicates columns of data to be treated as channel's output}

\item{side_variables}{is a character vector that indicates side variables' columns of data}
}
\value{
A character object that includes a standard formula syntax to use in algorithm for capacity calculation
}
\description{
Formula for logistic regression algorithm
}
\examples{
formula_generator(signal="signal",response="response", side_variables=NULL)
formula_generator(signal="inputX",response="responseY", side_variables="SV1")
formula_generator(signal="signalX",response=c("r_1","r_2","r_5"), side_variables="SV")
}
