% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_signal_transform.R
\name{func_signal_transform}
\alias{func_signal_transform}
\title{Initial verification and transformation of input variable}
\usage{
func_signal_transform(data, signal)
}
\arguments{
\item{data}{is a data.frame}

\item{signal}{is a character that indicates columns of data that include the labels of input}
}
\value{
A data.frame that is a copy of data provided with signal column transformed to factor class. 
If signal has been numeric initially, additional column is created "signal_RAW" that is an exact
copy of original column
}
\description{
Internal, auxillary functions
}
\examples{
data=data_example1
data1=SLEMI:::func_signal_transform(data,"signal")
data$signal=as.character(data$signal)
data2=SLEMI:::func_signal_transform(data,"signal")
data$signal=as.numeric(data$signal)
data3=SLEMI:::func_signal_transform(data,"signal")
}
\keyword{internal}
