% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_discr_pairwise.R
\name{prob_discr_pairwise}
\alias{prob_discr_pairwise}
\title{Probability of pairwise discrimination}
\usage{
prob_discr_pairwise(dataRaw, signal = "signal", response = "response",
  side_variables = NULL, formula_string = NULL, output_path = NULL,
  scale = TRUE, model_out = TRUE, lr_maxit = 1000, MaxNWts = 5000)
}
\arguments{
\item{signal}{is a character object that indicates columns of data to be treated as channel's input}

\item{response}{is a character vector that indicates columns of data to be treated as channel's output}

\item{side_variables}{is an optional character vector that indicates side variables' columns of data, if NULL no side variables are included}

\item{formula_string}{is a character object that includes a formula syntax to use in algorithm for capacity calculation}

\item{model_out}{is the logical indicating if the calculated logisitc regression model should be included in output list}

\item{lr_maxit}{is the number of iteration of iterative algorithm of logistic regression}

\item{data}{must be a data.frame object}

\item{maxNWts}{is the maximum acceptable number of weights in logistic regression algorithm - the higher, the slower the computations}
}
\value{
a list with three elements:
\itemize{
\item output$prob_matr - confusion matrix of logistic regression predictions
\item output$model     - optimal probability distribution
}
}
\description{
INPUT:
}
