% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapz_simple.R
\name{trapz_simple}
\alias{trapz_simple}
\title{Numerical intergration by trapezoidal rule}
\usage{
trapz_simple(x, y)
}
\arguments{
\item{x}{is a numeric vector}

\item{y}{is a numeric vector of the same length as x}
}
\value{
A numeric value of integral approximated by trapezoids
}
\description{
This function calculate the integral of function described by set of points (x,y) using trapezoidal rule.
}
\examples{
trapz_simple(x=1:10,y=exp(1:10))
}
\keyword{internal}
