% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_nfkb}
\alias{data_nfkb}
\title{Data from experiment with NFkB pathway}
\format{A data frame with 1500 rows and 3 variables:
\describe{
\item{signal}{Label of input}
\item{response}{The value of output}
\item{sideVar}{Label of the type of given observation}
}}
\source{
in-house experimental data
}
\usage{
data_nfkb
}
\description{
A dataset describing simple one dimensional intput - one dimensional output channel 
with 500 observations per input. Each observation can be in addition of 3 types
that occurs with propensities (0.6,0.3,0.1), resepctively. 
Conditional output distributions are Gaussian.
}
\keyword{datasets}
