% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLEMI.R
\docType{package}
\name{SLEMI}
\alias{SLEMI}
\alias{SLEMI-package}
\title{Statistical Learning based Estimation of Mutual Information (SLEMI). A package for estimating mutual information and channel capacity from experimental data.}
\description{
The package SLEMI is designed to estimate channel capacity between finite 
state input and multidimensional output from experimental data. For efficient computations, 
it uses iterative algorithm based on logistic regression.
}
